python_services/requirements.txt
# VoxMind Python Services - Requirements
# Python 3.10+

# Core ML
torch>=2.2.0
torchvision>=0.17.0

# NVIDIA NeMo Parakeet (moteur de transcription alternatif, CPU-friendly)
nemo_toolkit[asr]>=2.0.0

# PyAnnote pour identification des locuteurs
# Note: requiert un token HuggingFace (HUGGINGFACE_TOKEN env var)
# pyannote.audio 4.0.4 est la version actuelle stable (mars 2026)
# La spec mentionnait 3.1.0, mais 4.x est recommandé
pyannote.audio>=3.1.0

# Audio processing
numpy>=1.26.0
scipy>=1.12.0
librosa>=0.10.0
soundfile>=0.13.0
pedalboard>=0.5.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Whisper (optionnel, si whisper_server.py est utilisé)
openai-whisper>=20250625

# Utils
tqdm>=4.66.0
pandas>=2.1.0
